+++ backend/src/modules/catalog/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const { CatalogRepository } = require('./CatalogRepository');
+const { CatalogService, CATEGORIES } = require('./CatalogService');
+
+module.exports = { CatalogRepository, CatalogService, CATEGORIES };